requirements.txt
PyQt5>=5.14
pillow
opencv-python
numpy
ultralytics